+++ databruce/static/databruce/js/setlistTable.js
@@ -5,8 +5,8 @@ const SONG_BADGES = [
   { key: 'instrumental', class: 'badge-info', label: 'Inst.', title: 'Instrumental' },
   { key: 'sign_request', class: 'badge-info', label: 'Sign', title: 'Sign Request' },
   { key: 'nobruce', class: 'badge-warning', label: 'No Boss', title: 'Song Played Without Bruce' },
-  { key: 'debut', class: 'badge-primary d-lg-none', label: 'Tour Debut', title: 'Tour Debut' },
-  { key: 'premiere', class: 'badge-secondary d-lg-none', label: 'First', title: 'First Time Played' },
+  { key: 'debut', class: 'badge-primary d-inline-block d-md-none d-lg-none', label: 'Tour Debut', title: 'Tour Debut' },
+  { key: 'premiere', class: 'badge-secondary d-inline-block d-md-none d-lg-none', label: 'First', title: 'First Time Played' },
 ];
 
 function getTourPositionBadge(row) {
@@ -14,8 +14,8 @@ function getTourPositionBadge(row) {
 
   if (num && total && num === total) {
     return total === 1
-      ? createBadge('badge-secondary d-lg-none', 'Tour Only', 'Only Tour Performance')
-      : createBadge('badge-warning d-lg-none', `Tour Last`, `Final Tour Performance<br>(${num}/${total})`);
+      ? createBadge('badge-secondary d-inline-block d-md-none d-lg-none', 'Tour Only', 'Only Tour Performance')
+      : createBadge('badge-warning d-inline-block d-md-none d-lg-none', `Tour Last`, `Final Tour Performance<br>(${num}/${total})`);
   }
   return null;
 }